price cent (#103)

+++ Rij62/Models/api/ApiPutProduct.cs
@@ -4,7 +4,7 @@ public class ApiPutProduct
 {
     public required MultiLangString Title { get; set; }
     public required MultiLangString Description { get; set; }
-    public required int Price { get; set; }
+    public required decimal Price { get; set; }
     public required int Btw { get; set; }
     public required int Stock { get; set; }
     public required string ImgURL { get; set; }
